dom/events/test/test_bug1150308.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1150308
-->
<head>
  <title>Test for Bug 1150308</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<div id="host"><span id="distributeme">Foo</span></div>
<script type="application/javascript">

/** Test for Bug 1150308 **/
SimpleTest.waitForExplicitFinish();

SimpleTest.waitForFocus(function() {
  var host = document.getElementById("host");
  var shadow = host.createShadowRoot();
  shadow.innerHTML = '<style>.bar:active { color: rgb(0, 255, 0); }</style><div class="bar" id="inner"><content></content></div>';
  var inner = shadow.getElementById("inner");
  var distributed = document.getElementById("distributeme");

  is(window.getComputedStyle(inner).color, "rgb(0, 0, 0)", "The div inside the shadow root should not be active.");

  synthesizeMouseAtCenter(distributed, { type: "mousedown" });

  is(window.getComputedStyle(inner).color, "rgb(0, 255, 0)", "Div inside shadow root should be active.");

  synthesizeMouseAtCenter(distributed, { type: "mouseup" });

  is(window.getComputedStyle(inner).color, "rgb(0, 0, 0)", "Div inside shadow root should no longer be active.");

  SimpleTest.finish();
});

</script>
</body>
</html>
